heartbeat graph bug fix

+++ cat-home/src/main/java/com/dianping/cat/report/graph/ValueTranslater.java
@@ -1,7 +1,7 @@
 package com.dianping.cat.report.graph;
 
 public interface ValueTranslater {
-	public int getMaxValue(double[] values);
+	public double getMaxValue(double[] values);
 	
-	public int[] translate(int height, int maxValue, double[] values);
+	public int[] translate(int height, double maxValue, double[] values);
 }
